Put some better limits on slippymap tiles.
Even though we did not draw or attempt to create them,
we could try to iterate past some pretty large numbers
of tiles.  That made things slow.

There was also a bug when calculating very large tileset
sizes where the integer would overflow.  Make it a double
so that the calculation at least goes through



git-svn-id: https://svn.openstreetmap.org/applications/editors/josm@18590 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4