New autocentering technique, based on position of cursor on the map. User configures the tunable:
livegps.center_factor=80 (80% - default value)

It represents a value (in percent) of a rectangular screen
area where the cursor should be present. If cursor leaves the
area, centering is performed.

Time based centering can now be turned off by default with:

livegps.center_interval_msec=0

But is turned on by default for compatibility.


git-svn-id: https://svn.openstreetmap.org/applications/editors/josm@25809 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4